Fixed bugs and added better feedback
Fixed bug that counted the number of workouts completed improperly. Fixed a bug that was causing a text view to not update. Fixed a bug where the level started at 2 instead of one. Lastly, made it so that when a workout program is selected, it prints which one you selected to the screen in the workout tab.